pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "data-ingestion" 
version = "0.1.0"
authors = [
  { name="William Morrill", email="wkm2109@columbia.edu" },
]
description = "A Python library to ingest files into the data lake."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

dependencies = [
    "requests",
    "PyYAML",
]
